[CORRECT] int → Byte 타입 변경
PlantScore, PlantDict, 관련 dto 변경

변경 이유
- 각 필드 0-2 사이로 작은 값
- Bean validation에서 @NotNull 적용하기 위해

+++ src/main/java/com/be1/plant4you/domain/plant/PlantDict.java
@@ -30,19 +30,19 @@ public class PlantDict {
     private String name;
 
     @Column(nullable = false)
-    private int viabilityLevel;
+    private Byte viabilityLevel;
 
     @Column(nullable = false, length = 500)
     private String viabilityDesc;
 
     @Column(nullable = false)
-    private int waterFreqLevel;
+    private Byte waterFreqLevel;
 
     @Column(nullable = false, length = 500)
     private String waterFreqDesc;
 
     @Column(nullable = false)
-    private int sunAmtLevel;
+    private Byte sunAmtLevel;
 
     @Column(nullable = false, length = 500)
     private String sunAmtDesc;
